SECURITY: Validate term length in Special:NewProperty
(Note: the SECURITY is meant to mark this as a patch that is initially
deployed to Wikimedia production via the security deployment process,
but the actual security impact is minimal. We just didn’t want the bug
to be abused before it was fixed, hence the non-public deployment.)

Special:NewItem had some code to validate the label, description and
aliases of the new Item against the standard term validators (which
mainly means that the length should be limited). Special:NewProperty was
missing this code, meaning you could create Properties with labels of
arbitrary length. Fix that, and clean up the code in Special:NewItem as
well while we’re at it.

In Special:NewProperty, inject a TermValidatorFactory, so that we can
use it to create validators for the label, description and aliases, then
call those validators and use their result, using the same code style as
in Special:NewItem (after the cleanup, described below).

In Special:NewItem, inject a ValidatorErrorLocalizer. For questionable
reasons, the whole Validator area of the code uses Result and Error
classes that basically fulfill the same purpose as MediaWiki’s Status
(and Message), but are separate from it. Previously, Special:NewItem had
a custom method (createStatusFromValidatorError) to translate from one
style to the other, but we also have a service (ValidatorErrorLocalizer)
that does the same thing, in a much better way (preserving all errors
instead of just one, supporting non-string message arguments…), so let’s
just use that.

Also in Special:NewItem (and Special:NewProperty), change the code to
append to a common Status instance and return it at the end, instead of
returning a fresh Status at the first error encountered. The effect of
this is that, for inputs with more than one error, users are shown all
the errors at once; this isn’t required for T308659, but seems like a
generally desirable improvement to me.

By an incredible stroke of luck, both of the services we need to inject
(ValidatorErrorLocalizer and, for Special:NewProperty,
TermValidatorFactory) come alphabetically after all other services
injected into those special pages. This means that we can keep the
alphabetical order in the extension-repo.json wiring (which is enforced
by a unit test), and at the same time append the new services at the
end, which should let us deploy this change with no errors.
Specifically, the deploy order should be:

1. sync extension-repo.json
2. sync SpecialNewItem.php and SpecialNewProperty.php (in either order)
3. sync the tests

This way, the special page factory functions are first called with extra
arguments (which is allowed), and then by the time those functions
require the added arguments, extension-repo.json already provides them.
I’ve tested this locally.

(Finally, in SpecialNewItemTest, fix a typo in a mock method name, since
we’re copying the method to SpecialNewPropertyTest anyways.)

Bug: T308659
Change-Id: I8171bfef73e525d73efa60b407ce147130ea4742