REST: Add ItemDataRetriever interface and implementation
This adds the ItemDataRetriever and the corresponding implementation
using EntityLookup. It also moves the field constants from the
GetItemRequest object to the ItemData object, since they are now also
needed by ItemDataRetriever service, and accessing use case data from
a domain service would violate the onion architecture rules.

We could also create a mapping between the requested fields in the use
case and the ItemData fields, but that seems unnecessary at this point.

Bug: T308435
Bug: T308431
Change-Id: I864271b6b5470b78be5c19c5a0fc9168bf7ee298