Include the current page title in static mapframe image URL
Without this patch, Kartographer renders an unusable static map
thumbnail (which can be seen when browsing with NoScript, or in the
network requests for a page). The title is required in order for the
Kartotherian server to retrieve the synthesized mapdata.

This is fundamentally ugly, because we would like the value formatting
to be independent of the page that the value is used on, but
Kartographer / Kartotherian requires this information. The request
context title might not always be correct for all the ways that the
formatter might be used, but it’s the best we have. (It still doesn’t
work for historical revisions: we’d really need to supply the revision
ID of the page being parsed as well, but couldn’t find a way to access
it without explicitly passing it through the formatter. As it stands,
the static thumbnail in outdated revisions will no longer work if the
coordinate was changed in the meantime.)

Bug: T309695
Change-Id: I5b3d9dc604608f21cfd311647fad4ee84b7c5e3c