Change API tests from unit tests to integration tests
These tests need to be run as integration tests, with a test DB set up,
since they pull in some hook handlers (e.g. from CentralAuth) that
aren’t ready to be run as pure unit tests.

We already have an ApiFormatReferenceTest in integration/includes/Api/,
so call the new one ApiFormatReferenceUnitTest, explaining in the class
comment how it’s kind of “unit” but not fully.

Bug: T310255
Change-Id: Ia6c6ffddd2689718601c6295da8d2f1b2cf77e53