Add profile parameter to entity search APIs
The parameter is configured by a new setting mapping API parameter
values to internal EntitySearchHelper $profileContext values. Out of the
box, only a 'default' profile is supported.

As a *temporary* special case, the parameter is only made available if
more than one profile has been configured. This will let us easily
deploy the code for this feature in production without actually enabling
it yet, then later enable it with a configuration change. (Effectively,
the setting that configures the profiles temporarily doubles as a
feature flag for the API parameter.) This unusual behavior is not
permanent: later (change I4139767c3a), we *will* make the API parameter
available unconditionally, even if “default” is the only legal value.

The PARAM_HELP_MSG_PER_VALUE instructs MediaWiki to show a help message
for each possible value, with a message key based on the parameter
value. This patch defines that message for the 'default' value; wikis
that configure additional values should also define the message at the
same time (otherwise MediaWiki will gracefully fall back to a “no
description” message), e.g. by creating the corresponding page in the
MediaWiki: namespace. (For Wikidata, I think we’ll define these
additional messages in the WikimediaMessages extension.)

Note that we use the array_key_first() method, which is new in PHP 7.3;
it’s available in production (which is still on PHP 7.2) via
symfony/polyfill-php73, so add that to our composer.json requirements
(using the same version as in MediaWiki core and mediawiki/vendor.git,
but with a ^ operator to perhaps make future version bumps easier).

Bug: T307869
Change-Id: Icb1b3057bc23dddcb5787ab2801bfaa600065533
Depends-On: Icf7f9f58b11850a58506e267eb4e01acd73f3733