Add debug logging for item term storage
We want to see what is happening with the term storage that seems to be
behaving strangely and sometimes dropping term store updates (T309445).
Log messages to the temporary WikibaseTerms channel.

The deleteTermsOfEntity logging is “just in case”, but I don’t expect it
to be relevant for the bug we’re seeing, so it’s only logged when
scheduled (ItemHandler), not when it runs (ItemTermStoreWriterAdapter).

Bug: T311307
Change-Id: I42b46b9ad36e42ea3a8e6b1bb7ade55b9ceb3641
Depends-On: Ic0d8999e3e28a76a5968f04c571bda70af34dce2