REST: Don't use defaults in tests if not needed
Follow-up to I93170979a1de4044b65152ef8cb0158a3a2be6fc.

The default test doubles pattern [0] is useful when the class under test
has multiple service dependencies and multiple test cases which override
different dependencies. This test previously used a default service for
the legacy serializer, but never overrides it. In that case we can create
the dependencies at the same time as the instance under test and get rid
of the `legacySnackSerializer` instance variable on the test class.

This patch also makes some minor improvements to tie the test inputs
more explicitly to the service test doubles.

[0] https://www.entropywins.wtf/blog/2018/08/01/clean-architecture-usecase-tests/

Change-Id: Ic5f49a5aed5f24729f1f7652e3634bd742eb26bc