Prepare switching to WordLevelDiff in SiteLinkDiffView
Refactors the code without making any behavioral changes yet.
generateChangeOpHtml() is renamed to generateHtmlDiffTableRow(), as in
change If7211af17a; it now receives HTML as the parameters instead of
creating HTML based on the input. The difference between badge and site
link changes is now made at the beginning (in generateDiffOpHtml()), not
buried deeply inside getChangedLine().

Also, use `instanceof` checks to distinguish between types of diffs,
since we rely on the different classes anyways. (We don’t really need
AtomicDiffOp, and I could be convinced to change that check back to
->isAtomic(), but for the three specific classes, this seems like a
clear improvement to me.)

Bug: T303317
Change-Id: I60e8b1015c62bf6682e667544e30e92e385c2973