Use inverted values for the unexpectedUnconnectedPage page prop
This allows efficient sorting by ascending name space
and descending page id (new pages first).

In order to deploy this, the new legacy format migration
setting needs to be switched to false, then the
PopulateUnexpectedUnconnectedPagePageProp
maintenance script can be run. During this process
the special page will show incomplete results, but this
seems acceptable for the short duration in question.

Change-Id: Ifb569e719ba65d2d5a6fad202e87fcc6eb669793