Create FallbackLabelDescriptionLookupFactory service
This will replace the LanguageFallbackLabelDescriptionLookupFactory
service, though we’re not migrating any of its users in this change. The
difference between the two is that the new service creates lookups that
are guaranteed to also resolve redirects, which the old service doesn’t.

Currently, CachingFallbackLabelDescriptionLookup resolves redirects
while LanguageFallbackLabelDescriptionLookup doesn’t; document this fact
in the phpdoc of both classes. In the longer term, I could imagine this
functionality moving around as well – with the new factory, callers
don’t have to care about that.

Bug: T312222
Change-Id: Ia94c239869bc6525b70039030920ae7fe2f3be90