Fix HistoryEntityAction::getPageTitle() escaping
I got this wrong in change I3440f337f8 (commit 5ceacf8725). The page
title returned here (ultimately passed into OutputPage::setPageTitle())
is nominally HTML – OutputPage will take care of dropping “bad” tags
(like <script>) while allowing some “good” ones (like <sup>), and
setting the <title> to the same text with all tags removed. Therefore,
we should parse() the message, so that any wikitext in the message
source is processed (to be filtered again by OutputPage) – but mark the
label and entity ID as plain-text parameters, so that the message will
already HTML-escape all of their contents, including “good” tags
(because labels are supposed to be plain strings, not wikitext or HTML).

To be clear, change I3440f337f8 did not introduce a security
vulnerability, because OutputPage::setPageTitle() always makes the title
safe – it just unintentionally made it so that some HTML markup in
labels (the same kind of markup that’s allowed in {{DISPLAYTITLE:}}) was
now rendered instead of escaped.

Compared to the state before change I3440f337f8, the interface messages
wikibase-history-title-with-label + wikibase-history-title-without-label
can now contain some wikitext that will be parsed in the history title
(again, only the {{DISPLAYTITLE:}}-valid kind), which seems like a good
thing to me. Previously, due to using text(), tags like <b> or <sup> in
the message source would work, but markup like ''' wouldn’t. (Another
alternative would be to use escaped() instead of text() or parsed(), so
that *no* kind of markup is allowed in the message source.)

Change-Id: Ice545455183b0deb51f277bbfaf6218c6a043edb