Use FallbackLabelDescriptionLookupFactory in item entity types
This ensures redirects are resolved when formatting up an Item as an
entity (rather than an Item *ID* as a snak value – that formatting is
wired up elsewhere), e.g. in the wbformatentities API, or in the Lexeme
header. This is the direct fix for T305032.

Properties are not included, because there is no EntityRevisionLookup
for federated properties, which means the caching lookup can’t look up
the latest revision ID of a federated property ID to use in the cache
key, and so that whole implementation doesn’t work for federated
properties. It would be nice to fix this at some point, but for now,
properties don’t get caching here. (Properties cannot be redirects, so
that part doesn’t apply anyways.)

Bug: T305032
Bug: T312223
Change-Id: I8d9d7b027942738d6d6e70b3a4ce56c43a20c1f2