Use LanguageNameLookupFactory in entity types
The LanguageNameLookup service is deprecated because it relies on the
global user language – here, we get a Language passed in, so we can just
pass that into the LanguageNameLookupFactory and avoid the deprecation.

Noticed while looking at this code for change I8d9d7b0279, but not
actually related to T312223.

Change-Id: I243536d624f35f3970e2d1e8c1e354736ab67b62