Use FallbackLabelDescriptionLookupFactory in client hook handlers
These are used to show the currently used entities when looking at the
info of a client page and when editing it, respectively (assuming the
2017 wikitext editor [1] isn’t used). With this change, they’ll resolve
redirects, and also add caching, which I think might be significant
especially for the EditActionHookHandler.

In the tests, we need to return a FallbackLabelDescriptionLookup, not
just any LabelDescriptionLookup. This was already wrong for the previous
factory (which should return a LanguageFallbackLabelDescriptionLookup,
after all), it just wasn’t noticed because the return type of the old
factory is only in the phpdoc, not checked at runtime.

[1]: https://www.mediawiki.org/wiki/2017_wikitext_editor

Bug: T312223
Change-Id: I2d9991e5f15ad975f8583835635ca8841f1633f6