Introduce DispatchingFallbackLabelDescriptionLookup
Without this class, the CachingFallbackLabelDescriptionLookup returned
by FallbackLabelDescriptionLookupFactory cannot be used for federated
property IDs, because federated properties don’t have an entity revision
lookup, which is required for the caching lookup. This new lookup
implementation wraps two lookups, one that works for all entity types
(Language…) and one that only works for most (Caching…), calling the
right one depending on the entity type. The factory returns the new
lookup, so that it can be used with any entity ID.

This is hopefully a temporary workaround; eventually, there should be an
entity revision lookup for federated properties, and this class can be
removed again. But this might not happen in the near future.

Bug: T312791
Change-Id: Idf198614b49f7f60c33b4c2026976e07e6d583b2