Use FallbackLabelDescriptionLookupFactory in EntityIdLabelFormatterFactory
The formatter being created here eventually ends up in the
SiteLinksView, where it’s used for the title (tooltip) of sitelink
badges. It should be quite rare for a badge item to become a redirect,
but if it happens, it’s still better to show the label of the target
item than to just show the item ID.

Also, in WikibaseQualityConstraints, this factory is used to format
constraint reports.

Bug: T312223
Change-Id: I727f0c02ae0b2bce52aa46d775e5c06c93c1fed7