Use FallbackLabelDescriptionLookupFactory in SpecialSetSiteLink
Like the parent commit (change I727f0c02ae), this is used for badges,
showing the label of the target item if a configured badge item ever
becomes a redirect.

If you’re testing this locally and you have both the merge source and
target configured as badge items, you’ll notice that only one will be
shown in the special page with this change. This is ultimately a
limitation of HTMLMultiSelectField: the options are specified as an
array, with the label as the key and the value as the value, so multiple
options (items) with the same label override one another. (It follows
that this isn’t actually limited to redirects either – it will also
happen if two unrelated badge items happen to have the same label.)

Bug: T312223
Change-Id: I392c31ed27bd1f0a0d77e729ae9d148e015d3e69