Fix phpcs-modified scripts
I4fe50f79aaf9858e28027f0877df82726c3fbd83 introduced an issue that
caused them to fail when there were staged removed files. This issue
was already present in the original version of the script, but only
for unstaged deletions which may not be as common.

phpcs-modified and fix-modified now no longer run for untracked files,
which people didn't seem to care about anyway.

Change-Id: Icf391999186f47580799cec210b5b8d400320260