Remove WikibaseTerms debug logging again
We decided to suspend the investigation for now, and this is too much
logging to just keep when we’re not actively using it. It can be
resurrected later if needed.

This effectively reverts change I42b46b9ad3 (commit 09f5e27c85) and
change Idcaee75235 (commit 7d97066036), except that the square brackets
for the job parameters when pushing the CleanTermsIfUnusedJob in
DatabaseTermStoreWriterBase stay arranged differently (on the same line
as the round parentheses, instead of the whole array being on one line),
because I prefer them this way.

Bug: T313039
Change-Id: Iaea7e596b08f93ed5405bb418313065e86cca2d9