Use FallbackLabelDescriptionLookupFactory in SpecialListProperties
And refactor the class a bit so that it doesn’t need to have a
PrefetchingTermLookup and LanguageFallbackChainFactory injected.
Previously, it prefetched all the property IDs from the underlying term
lookup, but then asked the label+description lookup factory for a new
lookup with every row. Instead, call the factory only once (and prefetch
the property IDs at the same time), and then use that factory for every
row. This should be equivalent, but in my opinion it’s much nicer.

Bug: T312223
Change-Id: I0a66e8bfcd8f150d7228714b8d4562db7879e22e