Stop using SingleEntitySourceServicesFactory for term lookups
We don’t need to get an EntityRevisionLookup that’s specific to the
given entity source. Services in the service container are based on an
EntityRevisionLookup (and other services) that works for any entity
source, dispatching based on the entity ID as needed, so we can just use
the service container’s RedirectResolvingLatestRevisionLookup directly.

Bug: T277731
Change-Id: I0a837d8c0c2ce24396c6e9d05a6269c853571e32