Use FallbackLabelDescriptionLookupFactory in SidebarLinkBadgeDisplay
The labels being looked up here are eventually used for the title=
attribute of a sitelink; if a badge item is unexpectedly a redirect, we
want the label of the redirect target to be used.

If you want to test this locally, note that the badge item must be
included in $wgWBClientSettings['badgeClassNames'] (a mapping from item
ID to CSS class), otherwise it won’t be included in the title.

Bug: T312223
Change-Id: I8bdc1c45ec688f24d36830eb7dbbe4d17b30a227