Stop using LanguageFallbackLabelDescriptionLookupFactory in Scribunto_LuaWikibaseLibrary
This is a rare case where (I think) we don’t want to replace
LanguageFallbackLabelDescriptionLookupFactory with
FallbackLabelDescriptionLookupFactory. The created lookup is here being
used as an internal layer for a TermLookup, which resolves redirects and
adds caching; we don’t want the inner lookup to resolve redirects and
cache terms as well. So instead, we get rid of the old factory by
directly having the CachingFallbackBasedTermLookup create a
LanguageFallbackLabelDescriptionLookup, after injecting the services it
needs for that into it.

In the tests, we can no longer mock the lookup returned by the lookup
factory, but we can override the CachingFallbackBasedTermLookup’s
internal list of lookups, so that it never tries to create a new
LanguageFallbackLabelDescriptionLookup. This isn’t great, but I think
it’s acceptable, and better than keeping the old factory around
indefinitely.

Bug: T312224
Change-Id: I24d6e2a749032009b6c455ec29f356cdb880733f