Streamline DatabaseEntitySource constructor for performance
The original motivation for this patch was to use the native
array_column() where possible. This is done in the tests. But it turns
out it's not possible in DatabaseEntitySource because array_column()
drops the array keys.

I ended reworking DatabaseEntitySource anyway to avoid looping the
same array multiple times.

It's also possible to remove some of the assertions without loosing
anything. These error situations will still be reported. The argument
for using suppressWarnings() is: When configuration is so heavily
broken it's ok when it makes PHP spit out warnings. As long as the
configuration error is reported and easy to fix, which it still is
thanks to the remaining assertions.

Change-Id: I20aab9fdbe098e6e00d070ac1d961f8880cfca3d