Use FallbackLabelDescriptionLookupFactory in DataAccessSnakFormatterFactory
This does not affect the labels that are actually shown, but it affects
usage tracking, and reduces the number of unnecessarily tracked label
usages on any wiki whose content language is not English (and which has
label usages for redirected items at all).

For example, if the client wiki’s content language is German (de), and
item A has a German label, and item B redirects to item A, and entity C
(may be the same as item A, or even another entity type) has a statement
for property P pointing to item B, and a wikitext page W contains
{{#statements:P|from=C}} (with P and C replaced with the actual IDs) –
then that page W would previously have label usages for item A in German
and English (or more generally, for the whole fallback chain). With this
change, it only has a label usage for item A in German (or more
generally, for the part of the fallback chain that was actually used).

This is because the snak formatting for the {{#statements}} parser
function (and possibly for Lua as well, I think?) doesn’t quite do usage
tracking at the same time as formatting. It formats the snak using
regular formatters that don’t know anything about usage tracking and
work for all sorts of data types – items, quantities, times, whatever –
then goes through the snak that was just formatted, checks if it
contains entity IDs (as entity ID values, or items as quantity units),
and tracks label usages for those. Prior to this change, the regular
snak formatting was already resolving redirects, but the usage tracking
afterwards wasn’t resolving them, so it would think that the entity had
no label, which means the whole language fallback chain needs to be
tracked (since a label in any of those languages should be used if it
appears). Now that the usage tracking part also resolves redirects, it
knows which label was actually used, and can add usage tracking only for
the part of the language fallback chain that is necessary.

In the tests, since we have to touch those lines anyways, simplify the
mocks by using createMock() instead of getMockBuilder() +
disableOriginalConstructor() + getMock().

Bug: T312223
Change-Id: I4996517970cafa2732d1bfc8fcd815e694efbbf1