Hard-deprecate LanguageFallbackLabelDescriptionLookupFactory
We currently plan to remove it after the 1.39 release.

In the service wiring tests, use MWDebug to hide the deprecation
warning. To have access to the hideDeprecated() method, we’d have to
extend either MediaWikiUnitTest or MediaWikiIntegrationTest, but not all
service wiring tests work as proper unit tests yet (T277464), and they
definitely shouldn’t be integration tests. (We could get the method from
MediaWikiTestCaseTrait, but that trait isn’t stable to use.)

Bug: T312224
Change-Id: Ic699d16dede96639eadc9162db0517e1eed4778d