Don't allow a redirect badge on a sitelink to a redirect be removed
The specification of `wbsetsitelink` and `wbeditentity` allow to not
specify the target pagename of an existing sitelink when only editing
its badges.

However, to correctly resolve its redirects, we need to know the
pagename, so that we can resolve it depending on the badges.

Thus we need to enable the deserializer for `wbeditentity`, which is
doing the redirect-resolving, to know about the existing sitelinks.

This now changes somewhat fundamentally what information is available to
the ChangeOpDeserializers in that it adds the EntityId, if it exists.

An alternative, that was considered, was to push down resolving the
redirects to the step when the sitelink is applied.
This would have had the advantage that the redirect resolving logic
would have to happen only once for all three entry-points. However, this
is not practical because the ChangeOps are not created by a
factory/service and so it is not straight forward to inject
dependencies/lookups.

This approach is consistent with data value normalization, which also
happens as part of change op deserialization:

* ClaimsChangeOpDeserializer::createStatementChangeOp() calls
* StatementChangeOpFactory::newSetStatementOp() which calls
* StatementNormalizer::normalize().

An integration test will be added in Ie3ef9433

Bug: T314848
Change-Id: Ia8c808e3baf0499d17ea35a485e8214c45f4c44f