EntityLinkFormatterFactoryTest: Use mock for language
Using Language::factory creates various side effects, including the
circular dependency issue we are dealing with in GrowthExperiments. This
is very much a band-aid fix. (See T315489#8164239 for a step through of
the stacktrace.)

Though it is unclear why only the calls in this test are causing a
problem while the ones in the other tests are seemingly fine.

Bug: T315489
Change-Id: I953826bf0dcbe13f3d210c1b3fbb58f6acc6d306