Use SelectQueryBuilder in SqlEntityIdPager
I think it’s much more readable this way, where the code assembling the
query isn’t spread across three different methods anymore – even if it
means fetchIds() is a bit longer than before.

The select fields are simplified by removing the array_unique() +
array_merge(), and only selecting the LinkCache::getSelectFields(). I
don’t see a reason to select page_title and page_namespace at all (we
don’t use them in the result, as far as I can tell); page_id is one of
the LinkCache::getSelectFields(), and almost guaranteed to stay there as
well, so instead of adding page_id to the fields at runtime, we add a
test that checks the page_id gets included in the selection.

Bug: T312510
Change-Id: I38a72ff3215752a8ae20c609c1e3f569e6323d3e