1) Define the most basic entity
This entity just contains a string.
I am not implementing all of the needed things yet,
so that I can work through the creation process and
see the minimum set of required things.

So far this commit
a) adds a class that extends EntityDocument (required)
but there are ways this interface could be reduced
b) adds an ID, that is just a simple string (required)
again, this interface could be simplified
c) adds a special page so that I would be able to try and
create the entity.
Note: It would have been nicer to be able to say that an entity
type brings an interface with it, instead of needing to add this
to WikibaseRepo extension.json directly

Change-Id: Ia1750a56fd70d6a9699e592a6805ddb61bcf8428