Stop using prophecy in ItemPropertyIdHtmlLinkFormatterTest
Mostly a straightforward replacement of:

- prophecize() → createMock()
- someMethod()->willReturn() → method( 'someMethod')->willReturn()
  - someMethod( $arg ) → with( $arg )
- reveal() → nothing

However, there are a few further differences. In givenItemHasLabel() and
givenPropertyHasLabel(), we now need to pass the entity ID string into
the entity to be created, because otherwise some tests fail with the ID
being different – I guess ->getTitleForId( new ItemId( $itemId) ) didn’t
actually assert whether the method was called with $id or something
else, whereas ->with( new ItemId( $itemId ) ) includes that assertion?
And finally, prophecy allows return values to be changed later, but
PHPUnit doesn’t – therefore, the Title methods isLocal() and
isRedirect() can’t simply be mocked to return one boolean and then later
changed to return another boolean, but rather have to return a callback
that returns a boolean by reference, and then we change that reference
later (after also passing it by reference into the class we return).
That’s a bit unfortunate, but not the end of the world, I think.

Bug: T315966
Change-Id: I2955d7251f99a9cc76f78318b82aed25d772357e