Send language names to wikibase.getLanguageNameByCode()
For wikis without ULS, this may be the first time we’re providing
language names on the JS side at all: we can have the server-side code
send them as a generated file of the ResourceLoader module.

For wikis with ULS, the majority of the language names is already sent
to the client as part of a ULS module. We don’t want to send this data
twice, so we leave it out of the Wikibase data. The client-side code
looks for language names in the ULS data, as before; however, it now
only does that after first looking in the data we sent from Wikibase.

getLanguageNameByCode() is effectively rewritten; the ULS logic is the
same as before, but organized differently, so that you can very easily
see the fallback chain in the main function: all the other functions
have the same signature, taking a string language code and returning a
string language name or null.

This fixes the language name of 'mul', as well as many other language
codes, in the classical / legacy termbox, as well as in other places
using this function.

Bug: T261851
Bug: T311626
Change-Id: Ic61b6146c338e69336cf6bbbb4525c62dbe9bb90