Simplify ClaimSummaryBuilder::buildClaimSummary
This simplifies some code which looks like it used to handle multiple
statements at once and is now obsolete.

It also slightly changes the summary args to not wrap the Snak object
containing the statement's value in an array, because
ClaimSummaryBuilder is guaranteed to only contain a single statement.
This has no effect on the formatted summary, since a list with a single
element is formatted the same way as the value element itself.

Change-Id: I13978b4b0a34ca9fba6b9ce5204ad76b38236448