Make sure language lookups return plain arrays
array_unique() preserves array keys, so if it removes any elements then
the resulting array has holes in the keys.

    [ 'A', 'B', 'A', C' ] ⇒
    [ 'A', 'B', 3 => 'C' ]

Use array_values() to reindex the array to consecutive keys. This is
especially important when we’re going to send the array to JS, where
arrays with non-consecutive keys become objects.

Bug: T311617
Change-Id: Ia35d02c5cdaa173d514d73998300ad8b123a94cc