Send user preferred content languages to client-side
The client-side update of the termbox should be based on the same list
of languages that was used to render the termbox server-side in
OutputPageBeforeHTMLHookHandler::getLocallyRenderedEntityViewPlaceholderExpander,
i.e. a UserPreferredContentLanguagesLookup. Send this list as an
additional variable to JS.

We still need the original list of user-specified languages, though,
since this determines whether we should add additional languages based
on what ULS thinks the frequent languages are (based on GeoIP data).
These languages should only be added for users who have no babel (not on
their user page, or who aren’t logged in at all), not for users that
have any babel, even if their babel is just one element long. (Which
means we can’t tell the difference based on the new variable alone.)

The old code added the user language to the front of the list after
filtering the list for valid term language codes, which means that
?uselang=invalid will actually show “invalid” as a language code in the
JS-rendered term box. The new code keeps this behavior, but with a
comment that, frankly, I have no idea if this is intentional or not.

The JS tests are basically rewritten (though with a lot of the code
lifted from the old version), so a inline diff might not make much sense
there – it might be easier to just review the new file as a whole.

Bug: T311617
Change-Id: I5588b22378e0133ab91ee61fdc2ccb3d4a23efe3