Overhaul magic summaries documentation
The old documentation was quite outdated, but claimed to contain a
comprehensive list of all keys in Wikibase (which was definitely no
longer true). The new documentation is based on Thiemo’s excellent
commit message of change Ida48142866 (commit e195421ed8), extended with
more details and examples. I hope it’s accurate.

Also, fix the link to the documentation in AutoCommentFormatter, since
the file was moved a while ago. (The @ref syntax is turned into a nice
link by Doxygen; we use it in a few other classes as well.)

Change-Id: I8666b8a562f85c6f1abb454d5440482f337b3dd6