UnexpectedUnconnectedPagePrimer: Don't rely on affectedRows
After $dbw->replace(), $dbw->affectedRows() will
return the number of rows deleted plus the number
of rows inserted (thus, if we replace existing records
twice the number of pages affected), which is not what
we want to report to the user here.
We can do this as we only ever pass rows that
actually need to be changed to $dbw->replace().

Change-Id: I4dbec0bcbacf8f43fd73d8670059d1b7531f9ace