Add a way to extract content scoped search index data
The new method EntityHandler::getContentDataForSearchIndex() method is
meant for usecase when an Entity is being stored in a non-main slot
(e.f. MediaInfo).
ContentHandler::getDataForSearchIndex in Core does not know how to
deal with side slots and is only called on the main slot
ContentHandler.
This method which is scoped to the slot Content can re-used and/or
overridden for usecases when non-main slots are in use.  It might make
sense to make core aware of such situations but given that MediaInfo
is the sole usecase requiring this it might be a lot easier to limit
the scope of this change to Wikibase & WikibaseMediaInfo.

Bug: T190066
Change-Id: Ie4a7513910ff5272b54757c1a8aae1688bddb111