Re-add minimal stubs for ruflin/elastica library
These were removed, along with stubs for extensions, in commit
9c0cb2559a (change I9ea6a539f2), but while Zuul can clone other
extensions for us in CI, we need a stub for this library, at least on
release branches: mediawiki/vendor.git removes this library (and many
others) from release branches. (The stub probably wouldn’t hurt on
master either, though.)

The code is based on the old stubs, but much reduced, since barely any
code in Wikibase uses Elastica (most of it was moved to the separate
extension WikibaseCirrusSearch long ago).

Bug: T319390
Change-Id: Ib888a568a04ba916384e8f87230b3cbce7c9de55