Always use new "unexpectedUnconnectedPage" page prop format
Remove support for the legacy format, where the sort key was the
positive namespace; only support the new format, where it’s negative.

The only remaining mention of the legacy format (apart from other,
unrelated “legacy” things) is in UnexpectedUnconnectedPagePrimerTest,
where we keep the test for converting *from* the legacy format to the
current one: we still want to ensure that the maintenance script can do
this conversion, in case someone else used the legacy format.

Change-Id: I1588c547676d9d58f4288ebba9f3e684e34b37f2