Hard disable federated properties on REL1_37
For https://github.com/wmde/wikibase-release-pipeline/pull/357

It felt like the easiest place to do this was in the creation
of the repo settings array, as this is after user settings, but
before all access by Wikibase.

Other places were considered
 - on access of the 1 setting
 - Post extension load
But they were not chosen.

This also removes some docs from the branch and disables a bunch
tests that will fail with the forced setting to false..

Bug: T317890
Change-Id: Ie15df06ec559ae8a03e1b0580f96d3d087b8f162