Update data-values versions in composer.json
Revert "Lock data-values versions in composer.json" (commit 6d92f07baf,
change I8177e88920), and also declare compatibility with
data-values/number ^0.12.3, so that it’s possible to install Wikibase
with all the “old” versions of data-values (only supporting
Serializable) or with all the “new” versions (supporting __serialize()
and __unserialize()).

Also update the data-model tests for the new serializations. Note that
we also have tests for hash stability, and we *don’t* need to update
those: the hashes stay stable.

Bug: T301249
Change-Id: I8335800da3e7514f2b0126e09e009f7f00d353e6