REST: Handle patched statement value type mismatch
Respond with a more specific error if the type of the patched statement's datavalue doesn't match the datatype of the statement's property.

Only requires changes to repo/rest-api. Iterates over all the Snaks in the Statement and calls `SnakValidator::validate()`. This is similar to what the `SnakValidator::validateStatementSnaks()` method does.

Bug: T320659
Change-Id: I45800683d5923d77dd8575a9a16df9cd056822f3