Make a few tests more compatible with PHP 8.1
Implementing Serializable without implementing __serialize() and
__unserialize() is deprecated in PHP 8.1; the easiest way to avoid the
warning is to use a real DataValue class (which already implements the
new methods) instead of an uncustomized mock.

This also fixes one warning that, on the face of it, isn’t directly
related to Serializable:

> TypeError: PHPUnit\Framework\MockObject\InvocationHandler::__construct():
> Argument #1 ($configurableMethods) must be of type array, null given.

So this might also be considered part of T319219.

Bug: T301249
Change-Id: I2f4c896a17ca6f2a8594b857bfd2c5bfa2b81024