SitesModuleTest: Configure file paths for sites
SitesModule uses the api.php file path of the site, so we should
configure the file path; previously, this was the empty string (compare
the old test data) due to some type confusion, but MediaWiki change
I5ed7e169c9 will make this an error.

Also change the link path from just /$1 to /wiki/$1 so it doesn’t
overlap with the file path (/w/$1).

Change-Id: If20688411e3f0108b0c0ba3b91e5cc264b70a208
Needed-By: I5ed7e169c9753486c765fd816d1b21016c5c1def