Add strict types to LanguageNameLookup
This is in accordance with ADR 11. Also, it makes PHPUnit automatically
return a string from getName() instead of null (in tests that don’t mock
the method explicitly); this in turn fixes a PHP 8.1 deprecation warning
in EntityViewPlaceholderExpanderTest::newExpander().

Also add a TODO about the deprecated Language::fetchLanguageName()
method (T260340) since we’re already touching this code.

Bug: T319219
Change-Id: Id03517c3f2583eeb61d3d43eb35ee471974f4e9f