Use strict types in EntityRevision, RedirectRevision constructors
And remove tests for passing in values of the wrong types, which are
pointless now. The tests that used null as the $mwTimestamp produced
warnings under PHP 8.1 when that null was passed into preg_match().

WikiPageEntityRevisionLookupTest was actually using an int for the
timestamp, which is wrong not only for EntityRevision but also as the
mocked return value of RevisionRecord::getTimestamp(). Fix that.

Bug: T319219
Change-Id: Id3b6fc9d7f74dcdd4abd367cb0c2e6db05c7afdc