Use SettingsBuilder to register hook
Manipulating $wgHooks on the fly is deprecated and unreliable.

Note that we have to use registerHookHandlers() (allowed in “loading”
and “registration” stages), not registerHookHandler() (only allowed in
“loading” stage).

For the restored Phan suppression in Wikibase.searchindex.php, compare
change I7a59bf8e7b (commit a9d79567c7); in particular, it’s very
intentional that this file doesn’t use `global $wgHooks`. (I’m not 100%
sure whether that’s still important or it changed again, but let’s leave
it alone.)

Change-Id: Ic883ebf6963cb71da951a12492503832328ab16f
Depends-On: I334a103e02fd905faafc43c7c5b95996bc91fd18