REST: Replace ValidationError "source" with "code"
We haven't really used the "source" field as such, but rather as a
discriminator of validation error objects to map them to their
respective use case response objects. This patch now removes "source" in
favor of a new "code" field which better describes what it is. This is
inspired by Symfony's `ConstraintViolationInterface` [1].

It also moves the definition of these code values out of the use case
validators and into the more generic ones, where the context fields are
also defined.

[1]: https://github.com/symfony/validator/blob/6.1/ConstraintViolationInterface.php#L113

Bug: T322551
Change-Id: I9d2a1fb7d5c00ba1ad1e7d3b3590278571848e4e