Add strict types to StringNormalizer and tests
If someone passes null into StringNormalizer, we want that to throw an
error on all PHP versions, not just a warning on PHP 8.1.

Remove now-redundant phpdoc comments; some of them had a bit of text
beyond just the type, but IMHO the text was sufficiently redundant with
the main method documentation that it can still be removed.

Change-Id: I8f03bc4fee8cb06cdd43b3fd5070501ecbd9af98