Handle null/false $value in HTMLAliasesField
Treat it like the empty string (empty list of aliases); validate(),
below, may later throw an error for this if the aliases are required.

This should fix 10 of the 24 remaining errors when running
WikibaseLexeme CI under PHP 8.1.

The CI errors indicate that the actual $value was always null (never
false), so only change the param type to ?string, not to ?string|false.

Bug: T319219
Change-Id: I055d9fc352c4bccfe7cfc927866c64a1c707a90b